Split utils.js into multiple files; improve generateUsername()
