+++ package.json
@@ -25,7 +25,7 @@
     "cookie-parser": "^1.4.4",
     "dataloader": "^1.4.0",
     "dotenv": "^8.1.0",
-    "ejs": "^2.6.2",
+    "ejs": "^2.7.1",
     "express": "^4.17.1",
     "express-graphql": "^0.9.0",
     "firebase-admin": "^8.4.0",
@@ -69,7 +69,7 @@
     "chai": "^4.2.0",
     "eslint-config-prettier": "^6.1.0",
     "eslint-plugin-prettier": "^3.1.0",
-    "husky": "^3.0.4",
+    "husky": "^3.0.5",
     "lint-staged": "^9.2.5",
     "minimist": "^1.2.0",
     "prettier": "^1.18.2",
