+++ src/server/utils/index.js
@@ -1,0 +1,11 @@
+/**
+ * React Starter Kit for Firebase
+ * https://github.com/kriasoft/react-firebase-starter
+ * Copyright (c) 2015-present Kriasoft | MIT License
+ */
+
+export * from './map';
+export * from './relay';
+export * from './type';
+export * from './fields';
+export * from './username';
