+++ src/server/utils/relay.js
@@ -1,0 +1,19 @@
+/**
+ * React Starter Kit for Firebase
+ * https://github.com/kriasoft/react-firebase-starter
+ * Copyright (c) 2015-present Kriasoft | MIT License
+ */
+
+import { fromGlobalId as parse } from 'graphql-relay';
+
+export function fromGlobalId(globalId, expectedType) {
+  const { id, type } = parse(globalId);
+
+  if (expectedType && type !== expectedType) {
+    throw new Error(
+      `Expected an ID of type '${expectedType}' but got '${type}'.`,
+    );
+  }
+
+  return id;
+}
